Fix编译错误

+++ testsuite/core/helper/TestCase_Core_Helper_StlHelper.cpp
@@ -138,7 +138,7 @@ int TestCase_Core_Helper_StlHelper::Run(int argc, char *argv[])
     // array test.
     std::array<int *, 10> arr{nullptr};
     for (size_t i = 0; i < arr.size(); ++i)
-        arr[i] = new int(i);
+        arr[i] = new int(static_cast<int>(i));
     LLBC_STLHelper::DeleteContainer(arr);
     LLBC_PrintLn("after delete array, size: %lu, max_size: %lu", arr.size(), arr.max_size());
